+++ src/v1/sources/shopify/data/CheckoutStartedConfig.json
