+++ src/v1/sources/shopify/data/CheckoutStepCompletedConfig.json
