+++ src/v1/sources/shopify/data/CheckoutStepViewedConfig.json
