+++ src/v1/sources/shopify/data/ProductAddedOrRemovedConfig.json
