--- src/v1/sources/shopify/v0/data/mapping.json
@@ -1,10 +1,0 @@
-[
-  {
-    "sourceKeys": "line_items",
-    "destKeys": "products"
-  },
-  {
-    "sourceKeys": "id",
-    "destKeys": "properties.order_id"
-  }
-]
