--- src/v1/sources/shopify/v0/data/productMapping.json
@@ -1,18 +1,0 @@
-[
-  {
-    "sourceKeys": "id",
-    "destKey": "order_id"
-  },
-  {
-    "sourceKeys": "total_price",
-    "destKey": "value"
-  },
-  {
-    "sourceKeys": "total_tax",
-    "destKey": "tax"
-  },
-  {
-    "sourceKeys": "currency",
-    "destKey": "currency"
-  }
-]
