--- src/v1/sources/shopify/v1/data/identifyMapping.json
@@ -1,112 +1,0 @@
-[
-  {
-    "sourceKeys": "id",
-    "destKeys": "userId"
-  },
-  {
-    "sourceKeys": "email",
-    "destKeys": "traits.email"
-  },
-
-  {
-    "sourceKeys": "first_name",
-    "destKeys": "traits.firstName"
-  },
-
-  {
-    "sourceKeys": "last_name",
-    "destKeys": "traits.lastName"
-  },
-
-  {
-    "sourceKeys": "phone",
-    "destKeys": "traits.phone"
-  },
-
-  {
-    "sourceKeys": "addresses",
-    "destKeys": "traits.addressList"
-  },
-
-  {
-    "sourceKeys": "default_address",
-    "destKeys": "traits.address"
-  },
-
-  {
-    "sourceKeys": "shipping_address",
-    "destKeys": "traits.shippingAddress"
-  },
-
-  {
-    "sourceKeys": "billing_address",
-    "destKeys": "traits.billingAddress"
-  },
-
-  {
-    "sourceKeys": "accepts_marketing",
-    "destKeys": "traits.acceptsMarketing"
-  },
-
-  {
-    "sourceKeys": "orders_count",
-    "destKeys": "traits.orderCount"
-  },
-
-  {
-    "sourceKeys": "state",
-    "destKeys": "traits.state"
-  },
-  {
-    "sourceKeys": "total_spent",
-    "destKeys": "traits.totalSpent"
-  },
-  {
-    "sourceKeys": "note",
-    "destKeys": "traits.note"
-  },
-  {
-    "sourceKeys": "verified_email",
-    "destKeys": "traits.verifiedEmail"
-  },
-  {
-    "sourceKeys": "multipass_identifier",
-    "destKeys": "traits.multipassIdentifier"
-  },
-  {
-    "sourceKeys": "tax_exempt",
-    "destKeys": "traits.taxExempt"
-  },
-  {
-    "sourceKeys": "tags",
-    "destKeys": "traits.tags"
-  },
-  {
-    "sourceKeys": "last_order_name",
-    "destKeys": "traits.lastOrderName"
-  },
-  {
-    "sourceKeys": "currency",
-    "destKeys": "traits.currency"
-  },
-  {
-    "sourceKeys": "marketing_opt_in_level",
-    "destKeys": "traits.marketingOptInLevel"
-  },
-  {
-    "sourceKeys": "tax_exemptions",
-    "destKeys": "traits.taxExemptions"
-  },
-  {
-    "sourceKeys": "sms_marketing_consent",
-    "destKeys": "traits.smsMarketingConsent"
-  },
-  {
-    "sourceKeys": "admin_graphql_api_id",
-    "destKeys": "traits.adminGraphqlApiId"
-  },
-  {
-    "sourceKeys": "accepts_marketing_updated_at",
-    "destKeys": "traits.acceptsMarketingUpdatedAt"
-  }
-]
